requirements.txt
nibabel
numpy
Pillow
tkinterdnd2
matplotlib
